Refactoring, launch URL to the GitHub repository
